# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NADITOS API
  description: |
    National Digital Transport Operating System — public HTTP API.

    Every request flows through the API gateway, which:
      - injects an `X-Request-Id` header,
      - verifies the bearer JWT for protected routes,
      - enforces per-route role gates and per-(tenant, route) rate limits.

    Tenant identity is carried in the JWT claim `tid`. Clients may send
    `X-Tenant-Id` for routes that don't require auth (login); for
    authenticated routes the gateway rejects mismatches.
  version: 1.0.0
servers:
  - url: https://api.naditos.example
    description: Production gateway
  - url: http://localhost:8080
    description: Local docker-compose
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /v1/audit/events:
    get:
      summary: List recent audit events for a tenant
      parameters: [{ name: tenant_id, in: query, schema: { type: string } }]
      responses: { '200': { description: ok } }
...